Move CGGMP'21 entities to its submodule
