+++ synedrion/src/bench_internals.rs
@@ -1,4 +1,3 @@
 //! Public exports for use in benchmarks.
 
-pub use crate::cggmp21::benches;
-pub use crate::common::{KeyShare, PresigningData};
+pub use crate::cggmp21::{benches, KeyShare, PresigningData};
